requirements.txt
Django==1.10
djangorestframework==3.6.4
drf-tracking==1.3.1